+++ Student/StudentMain.java
@@ -1,0 +1,7 @@
+package Student;
+
+import java.util.Scanner;
+
+public class StudentMain {
+    
+}
