+++ Student/StudentVO.java
@@ -1,0 +1,12 @@
+package Student;
+
+public class StudentVO {
+
+    public void setMathScore(int mathScore) {
+        this.mathScore = mathScore;
+    }
+
+    public int getKorScore() {
+        return korScore;
+    }
+}
